Add ordering preservation and 64‑bit safety fixes
* Pass `ordering_` to Matrix constructors in binary ops.
* Use explicit `uint64_t` casts for element counts and assertions.
* Update copy constructor to allocate with casted size.

These changes ensure correct ordering is kept when creating new matrices and prevent overflow in index calculations.

+++ src/matrix.hpp
@@ -146,7 +146,7 @@ public:
                             ordering_{M.ordering_}
     {
         if( M.top_ != nullptr) {
-            std::copy((M.top_).get(), (M.top_).get() + M.m_ * M.n_, top_.get());
+            std::copy((M.top_).get(), (M.top_).get() + static_cast<uint64_t>(M.m_) * M.n_, top_.get());
         }
     }
